backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
click==8.1.6
dnspython==2.4.1
ecdsa==0.18.0
email-validator==2.0.0.post2
envparse==0.2.0
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
idna==3.4
passlib==1.7.4
pyasn1==0.5.0
pydantic==2.0.3
pydantic_core==2.3.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
